Using np.minimum and np.maximum to compute power_to_utilization.
This assumes that the minimum power is the actual minimum, according to
the input.
This may generate false values, if the system has the capability of turning
the GPUs in the system off (Assumed minimum when running is 75W, while
the the GPUs actually cosume 0 or 8W (standby)
This should report 0 regardless, but our remining calucluation should
account for this.

In the meantime this fixes the issue of negative utilization.